Model Training Completed

+++ src/components/data_ingestion.py
@@ -11,11 +11,14 @@ from sklearn.model_selection import train_test_split
 from src.components.data_transformation import DataTransformation
 from src.components.data_transformation import DataTransformationConfig 
 
+from src.components.model_trainer import ModelTrainerConfig
+from src.components.model_trainer import ModelTrainer
+
 @dataclass
 class DataIngestionConfig():
-    train_data_path = os.path.join('artifacts','train.csv')
-    test_data_path = os.path.join('artifacts','test.csv')
-    raw_data_path = os.path.join('artifacts','raw_data.csv')
+    train_data_path: str  = os.path.join('artifacts','train.csv')
+    test_data_path: str = os.path.join('artifacts','test.csv')
+    raw_data_path : str = os.path.join('artifacts','data.csv')
 
 
 class DataIngestion():
@@ -25,14 +28,18 @@ class DataIngestion():
     def initiate_data_ingestion(self):
         try:
             logging.info("Starting Data Ingestion Process...")
-            data_path = os.path.join("notebook","data","data.csv")
+            data_path = os.path.join('notebook/data/stud.csv')
+            if not os.path.exists(data_path):
+                raise FileNotFoundError(f"Data file not found at {data_path}")
             df = pd.read_csv(data_path)
+        
             logging.info(f"Data Ingestion Completed with shape {df.shape}")
             
             os.makedirs(os.path.dirname(self.config.raw_data_path),exist_ok=True)
             df.to_csv(self.config.raw_data_path,index=False)
             
             logging.info("Performing Train Test Split")
+            
             train_data,test_data = train_test_split(df,test_size=0.2,random_state=42)
             
             train_data.to_csv(self.config.train_data_path,index=False,header=True)
@@ -50,4 +57,8 @@ if __name__=="__main__":
     train_data,test_data = obj.initiate_data_ingestion()
     
     data_transformation = DataTransformation()
-    data_transformation.initiate_data_transformation(train_data, test_data)
+    train_arr,test_arr,_ = data_transformation.initiate_data_transformation(train_data, test_data)
+    
+    modeltrainer = ModelTrainer()
+    print(modeltrainer.initiate_model_trainer(train_arr,test_arr))
+
